Problem: workflow orchestration
I need a good way to expose executable, easily customizable business
workflows for one of my projects.

I don't really want to use something like Zeebe (https://zeebe.io/) or
Camunda (https://camunda.com/) because I fear that their are fairly
heavyweight systems and just managing their installations (Zeebe works
over Kubernetes) is too much for a one-man show. Managing Kubernetes
alone is a full-time job (or so I've heard).

Solution: implement my own lightweight workflow execution engine
in Elixir.

The hope here is that I can potentially achieve a lot with a moderate
investment in developing this engine. A low overhead platform that won't
need a lot of computing resources and will suffice for my needs.

I've tried to use BPE https://bpe.n2o.dev/ but it (seemingly) proved to
be harder to adapt to my needs.

Firstly, it's not trivial to follow the code. It seems to be heavily
tailored towards author's needs (which is fair) but not exactly mine. I
can even see some parts of this project are specifically there to serve
author's purposes (or at least it seems to be that way reading some
scarce comments).

It relies on local key/value storage for persistence which is not really
great for me (I use a standard Phoenix/PostgreSQL setup for that project
of mine). While KVS (underlying storage library for BPE) supports custom
backends, it still seems that it might not be suited for multi-tenant
deployments.

I might be wrong, but I've spent a number of hours trying to work around
its quirks and it left me unsatisfied.

I've also found Hashiru Bpmn Engine
(https://github.com/around25/hashiru-bpmn) but it seems to be unfinished
and abandonded as of this time.

So I decided to try and implement a usable subset of BPMN in Elixir. Not
being creative at the time of the start of the experiment, I named it
BPEXE (something along the lines of business process execution engine /
Elixir...)

I've tried to keep the initial design relatively straightforward and
avoid early optimizations. I am also trying to make it somewhat more
configurable than BPE.

Current idea is to have every component of the workflow work as a process
and communicate through messages. It might prove to be inefficient
later, and that's when I'll try to improve it. My current feeling is
that for the problems I am interested in, this should do just fine.

This commit is the initial draft and it is by no means complete in any
way. It was just able to do some minimalistic examples so it's something
worth publishing (I guess). Also, I've started some ideas in this draft
but didn't complete them. At this point, you just have to deal with this
:)

P.S. I am still hoping I will learn something useful from BPE. Kudos to
Maxim for implementing it!