Problem: configurability of parallel gateway
I might have not been too thorough reading BPMN specification, but I
thought that parallel gateway is quite limited in what it can do.

In particular, it seems to only join if all incoming branches send a
token.

Solution: add `bpexe:joinThreshold` attribute to `parallelGateway`

This will allow the parallel gateway to continue when a minimum number
("threshold") of incoming branches send a token. The result of the
results are discarded.

This changes the XML parser slightly to accomodate for namespaced
attributes.