Problem: instances can't survive restarts
This isn't great because if something goes wrong (a bug, hardware
meltdown, etc.) the state of the processes is lost and will have to be
restarted from scratch.

Solution: introduce a concept of flow handlers and instance initializer

Flow handlers are specifically designed to observe state changes in
various components as well as transactional boundaries. This allows us
to keep state stored (and restored) in case of restarts.

This also allows us to introduce another tool for debugging/tracing: a
"stepper" (can you think of a better name?). It allows to execute
workflows step by step and observe state changes.

This is different from observing log messages as those are rather
informational.

This patch is an early version that just seems to work enough to be
committed. I am not 100% proud of it (particularly the `__invisible__`
kludge -- not sure what to do about it, or whether I should do anything
about it).

I am pretty sure there's a lot of shortcomings in this approach but I
needed something to work with. In fact, I *do* know that it's somewhat
buggy.