Problem: transactional boundaries of actions don't work
I knew something wasn't right with the previous commit. I observe that
when I step through the workflow, things don't happen atomically, which
means there's a good chance data will be lost.

Solution: redesign sequence flows

They were processes and this presented an additional point of failure
and a challenge to synchronize its state with the state of the
neighbours.

This change moves functionality of sequence flows to `FlowNode`s, which
is, frankly, where it belongs. This is also making conditional sequences
flows implementation easier in the future (or so I hope) by keeping it
right in the same node that makes decisions based on them.

There is still a lot that I don't like about the implementation details,
particularly how this works with event-based gateways and events
(there's this whole business of handling non-acking of "held" messages,
sending back "invisible" messages and acking manually), but it seems to
work (somewhat, I know where it falls short). I have an idea with
regards to redesigning this, though. It'll like be synthesizing
additional gate logic for event management so that the
flow is always forward, and this whole funny business of "invisible"
messages and manual acking will no longer be necessary. I hope.