Problem: lack of exclusive gateway primitive
This is a very common one.

Solution: implement conditionExpression for flow nodes
and on top of that, build exclusive gateway.

This change forced `FlowNode` to account for the order of sequence flows
in order to establish priority. I think I also would like to add a
custom option to specify priority as a number.

`ExclusiveGateway` uses a bit of an uncoventional tactic of overriding
`send/3` and `send_message/3` and flow node state monitoring to intercept
the fact that a sequence flow was chosen.