Problem: references are not good message IDs
They won't effectively survive VM restart. Even though we can serialize
them to binaries, once we restore them in a new VM, there may be
collisions.

Solution: make ID generator configurable
and use CUID as a default choice.

I am not sure CUID is the best choice or whether XCUID is a good library
(haven't had a chance to read through it), but I wanted to make this
change to indicate the problem, solve it, and allow other ID generators
to be used.