Problem: no support for inclusive gateways
These are one of the major types describes in BPMN 2.0

Solution: implement sensor and inclusive gateways

I was thinking what would be the best way for a joining inclusive
gateway to know if particular sequence flows emanating from the forking
inclusive gateways were truthy or falsy.

First task was to find that fork. I am not yet 100% sure I've done it
right, but I've basically built a graph of predecessors and stopped at
finding another inclusive gateway, then topologically sorting the graph.
This probably needs some further improvements.

Then I considered a hack of asking that node's process which sequence flows
were truthy and which weren't. But that was hacky and meant storing more
of this information in FlowNode.

Instead, I decided to do a synthetic flow assembly again. The idea was
to have a "sensor gateway" which essentially has an additional incoming
flow that signals completion and an additional outgoing flow that
signals which other sequence flows were fired. The rest of flows are
effectively pass-through, cut into the workflow synthetically.