Problem: inclusive gateway will complete too early
I found a case where inclusive gateway will complete even though one of
the incoming flows will not reach it. In my test, that would be an
intermediateCatchEvent that never gets the signal.

Solution: fix a bug in the implementation

It was erroneously using an index of zero all times, always matching on
availability of the first incoming flow. I caught it when getting rid of
warnings (:facepalm). Do check your warnings often!