Problem: message `__txn__` is global
This mean is that if we send more than one message into the same
instance, they will get entangled in the ordering of transactions.
Since every message starts with 0, there will be collisions and order
confusion.

Solution: introduce a concept of "activation"

While I am not 100% clear on a clear definition of what activation is,
something like it should be introduced to allow for more than one active
interaction in the instance.

It's basically an independent flow (session?) in the process and flow handlers can
see this grouping and not confuse the transactions.

This change will definitely change how flow nodes will operate -- now
they will be able to get message's activation ID and distinguish
messages in different activations.

This is not yet covered well with tests (well, the multi-activation
aspect isn't covered at all...) but it's a start.