Problem: use of "transaction" (`txn`) term
I don't like the term I use to group state updates. It's an atomic counter that
allows us to group and order. But it's not an ID of some transaction. I
guess I called it a transaction because the behaviour of the system was
transactional in a sense -- with staging and committing data.

Solution: rename it to "generation"

I think it serves it a little better to show in what 'generation' of the
activation was the update.