Problem: use of "Message" term
It's a super-generic term and was kind of derived from the fact that we
send these entities as messages between flow nodes. But it's coming into
conflict with BPMN specification that also talks about messages in the
context of Receive Message / Send Message.

Solution: rename it to "Token" and change "content" to "payload"

It's unclear if we should keep the payload with the token, but I kind of
like the idea of passing data through and I think it should become
rather a dictionary of values that can be used ins scripts so that
there's always this context travelling with the token.