Problem: token's payload is rather useless
It's just a `nil` travelling throw the workflow. The original idea was
to let it carry some data down the flow.

Solution: change payload to be a Map and give scripts access to it

Now, both script tasks and sequence flow condition expressions have
access to `token` dictionary and a deep merge is used when tokens are
joined.