Problem: lack of serviceTask
This is arguably a very important component, but since the specification
is kind of vague on how particular service implementations are
specified, it took some time to think through.

Solution: add serviceTask handling and JSON support

In order to pass arguments to service, I needed some kind of structure
to pass around. Camunda seems to have its extension elements structured
around fields and their value. It didn't feel right to replice it.

Instead, I decided to opt for the low common denominator of JSON to pass
arguments to a service. Both plain JSON in `<bpxe:json></bpxe:json>` and
XSL-JSON (https://www.w3.org/2013/XSL/json/)
are supported.

Next, `BPXE.Service` was introduced (as a lightweight GenServer wrapper)
and Blueprint got some functionality to register services per Blueprint
PID. Upon registration, service links itself to the blueprint so that it
go down if blueprint goes down.

This change also introduces custom attribute extensions `bpxe:name` for
specifying service name in `serviceTask` and `bpxe:resultVariable` which
specifies which variable in token's payload should the result be stored
in (I am still thinking whether limiting it to token is a good idea)