Problem: BPMN's extensionElements handler loses scope
When handling BPMN definitions, we're keeping track of the nodes we add
things to. I noticed that when processing `extensionElements`, things
that were added later in the file, were not added to the right nodes.

Solution: don't drop the current node

There was apparently an innocently looking bug in the handling function.
It erroneously forgot to include current node in the scope when adding
extension elements node.