Problem: Lua is not a good fit for condition expressions
Firstly, one has to use `return <expr>`. This is a minor irritation.

Secondly, I think Lua is way too powerful for what condition expressions
should be -- they are simple boolean expressions and should not be able
to do anything complex. If they are, something is wrong.

Solution: use JMESPath instead of Lua in condition expressions

I've used JMESPath on numerous occassions before, and whille I don't
like some of its syntax choices, it works and it has a specification,
and more importantly, it's suited for data querying.

In my past experience, I was able to easily create parsers for other,
more "conventionally looking" expression languages on top of JMESPath --
that is, effectively translating them into JMESPath. I might do this
later on, but for now I'll just stick with it as it's an easy and safe
change that helps moving things in the right direction.