Problem: Lua is not a good fit for interpolating JSON payloads
We do this when we try to interpolate JSON payloads we send to services.
But just like in e94141e, it's not a good fit. The syntax is weird and
these expressions shouldn't do much beyond basic data manipulation.

Solution: replace it with JMESPath as well

(And yes, just like with condition expressions, an additional or
replacement syntax might come eventually, but this one gets the job done
easily and quickly)