Problem: storing token atomic counter in the counter itself
While this works reasonably well during the lifetime of the VM,
should these tokens be serialized and restored on a different VM, we
would require to go through all tokens stored in state snapshots to
reset their counters. This doesn't seem like a pleasant problem to
solve.

Solution: use persistent terms to access the counter

The idea is that we define an entity called Activation (we had it
before, but it didn't have its own struct) that refers to Blueprint ID,
Process ID and has its own serializable ID.

The activation does not change during its lifetime. However, we use
Erlang's persistent term storage to associate an atomic counter with a
[uniquely identifiable] activation. This way tokens can stay unchanged
upon state restoration.

Upon restoring the system, flow handlers can take an advantage of this
system to reset the token generation counter (which ETS one does now).

With this change, the silly business of using two 64-bit numbers to
count token generations is gone. It's erroneous, difficult to use and I
am ashamed I did it :) Fortunately, 2^64 generations is plenty.