Problem: blueprint test's flush_token/0 naming
Not clear why we need to "flush tokens". What does it mean?

Solution: rename it to "flush messages"

This is the result of my previous global search-and-replace when I
renamed messages to tokens in BPXE parlance, and I didn't catch this.

No harm done, but at least the meaning is clearer now.