Problem: state organization in the engine components
It really bothers me that the state of flow nodes and other components
is mixed with that of underlying layers like FlowNode or Base. It's
making code more difficult to navigate and easier to make mistakes.

Solution: reorganize state in layers

Subject components will now have their state at the top of the struct,
and all "sub-layers" will store their state under
`__layers__[layer_name]`.

It makes code a tad more verbose (or may be not even a tad?) but it's
definitely a step into a right direction and hopefully the verbosity of
this solution can be improved down the road.