Problem: Blueprint ID is not being auto-generated
If not passed as an option, it remains at `nil`

Solution: fix the bug in config initialization

That `Map.update` would not have worked here because of course the
struct has the key, so the update always did nothing. I didn't catch
this because there was no test for this. Now there is.