Problem: test runs are a bit slow
They are about 11 seconds currently on my computer. This is mainly
because of the use of `refute_receive` which naturally has to wait
until its timeout to ensure there was no such message.

Solution: run tests in parallel

I've tried a more sophisticated approach to this problem -- namely
devising an atomic token consumption counter. The counter itself worked
but I haven't quite managed to make the whole thing work reliably, and
it didn't cover all the cases of `refute_receive` because in some cases
we're testing in conditions in which we would not consider token to be
consumed.

So I've decided to stash away that experiment and simply try running
tests in parallel. It went down from ~11 seconds to ~2 seconds.

This is not to say that we don't need the token consumption counter. I
think this is going to be a valuable addition that gives further insight
into the process. But it'll have to come at a later point.