Problem: `Blueprint` and `instantiate_process` naming
These names are still bothering me. First one was fairly arbitrarily
chosen and although it kind of explains the essence of its nature, it is
nowhere to be found in BPMN's specification. This is not great.

The second one is confusing even me. In BPMN specification, they are
talking about different ways a process can be instantiated (start
events, event gateways, receive tasks). In BPXE's model instantiation
means provisioning actual nodes that handle the flow. Without it,
there's nothing working.

Yet, event gateways and receive tasks effectively imply that (at least
in BPXE's parlance) these processes are instantiated and await for
certain trigerring conditions to arrive.

Solution: rename `Blueprint` to `Model` and `instantiate` to `provision`

The reason I chose Model is that at least it bear some resemblance to
BPMN's specification. I know it's a word with overloaded meanings.

As for processes, there is no such thing as "process provisioning" in
specification, but at least it kind of makes sense in a conversational
context and does not clash with "instantiation" as per specification.

What this effectively means is that a "provisioned" process is
operational yet it is not "instantiated" or activated in any sense. I
think the next step would be ensuring that instantiation triggers are
activated upon provisioning (with or without them, processes can still
be started using a start event)