Problem: convoluted Engine.Process.flow_nodes/0 implementation
It's difficult to explain and maintain it. It relies on the fact that
flow nodes are linked to the process, that `Base` puts module name in
the process dictionary, etc.

Solution: simply collect flow nodes upon their creation

Implementing this change uncovered a bug: event-based gateway synthesis
inadvertently relied (and worked) on the order in which `flow_nodes/0`
returned these nodes. Incidentally, they seemed to have been returned in
the order they were created. However, the change as implemented collects
flow nodes in reverse order. One of the tests was failing.

The reason it was failing was that it was re-wiring nodes and re-adding
incoming and outgoing flows and since they were now collected in reverse
order, that's how they were added. Order is important for the
precedence gateway.