Problem: XML schema does not validate
Solution: import bpxe schema into Sematic.xsd

I wasn't able to find a better solution (but I am no XML guru!) but with a
small amendment to Semantic.xsd I think I was able to manage it. I essentially
added bpxe namespace to it:

```diff
--- Semantic.xsd.old    2020-12-24 11:13:43.000000000 -0800
+++ Semantic.xsd        2020-12-24 11:16:20.000000000 -0800
@@ -2,7 +2,10 @@
 <xsd:schema elementFormDefault="qualified" attributeFormDefault="unqualified"
        xmlns="http://www.omg.org/spec/BPMN/20100524/MODEL"
        xmlns:xsd="http://www.w3.org/2001/XMLSchema"
-       targetNamespace="http://www.omg.org/spec/BPMN/20100524/MODEL">
+       targetNamespace="http://www.omg.org/spec/BPMN/20100524/MODEL"
+       xmlns:bpxe="http://bpxe.org/spec/current">
+
+       <xsd:import namespace="http://bpxe.org/spec/current" schemaLocation="bpxe.xsd"/>

        <xsd:element name="activity" type="tActivity"/>
        <xsd:complexType name="tActivity" abstract="true">
```

Then I was able to override BPMN 2.0 schema with bpmn.xsd and validated BPMN
file with BPXE's extensions.

In order to do the validation, one needs to add this to the document root:

```xml
  xmlns:xsi="https://www.w3.org/2001/XMLSchema-instance"
  xsi:schemaLocation="http://www.omg.org/spec/BPMN/20100524/MODEL priv/schemas/bpmn.xsd http://bpxe.org/spec/current priv/schemas/bpxe.xsd"
```

Location of schemas can and should be changed to suit your needs. I plan to
upload these to bpxe.org soon, so it'll get a bit better.