Problem: syn usage is all over the place
`syn` is used for registry, meta-information and pub/sub functionality
all over BPXE. It does its job well, but it's difficult to get a
coherent picture of how it's being used (which naming patterns are
being used, which pub/sub groups are used for what) or replace it with
something else in the future, should there be a need.

Solution: centralize registry and channel in BPXE.{Channel,Registry}

This way the underlying implementation can be changed a little easier.

I alsoo added two Mix tasks, `bpxe.channel.list` and `bpxe.registry`
to provide a comprehensive view of names used for both channels and
registrations.

Closes #2