Problem: no support for properties
This makes tracking use of variables difficult as they have to be
extracted from scripts and whatever else can add them.

Solution: implement rudimentary handling of `property` element

This adds `bpxe:flow` boolean attribute to property to indicate that the
value of the property will be stored in the token and not the node.

This add `Process.get_property` API but I think it'll have be improved
in the near future when other item aware elements (such as `dataObject`
and/or `dataStore`) will be implemented as we'll need to refer solely by
element ID without knowing what kind of element it is. But I think it's
good enough for now.

Addresses #7