Problem: BPXE.Engine.BPMN API ergonomics
Having to supply a map with attributes for every node addition is just
adding noise for no gain.

Solution: allow passing attributes as keyword lists

This was an interesting thing to do.

Firstly, it highlighted the fact that in certain cases we use special
atom-typed attributes (`:body` or `:conditionExpression`). While it's
probably desirable to get rid of those, I didn't want to go that deep
and I needed to make sure only valid attributes are stringified from the
given keyword list. This was done by generating code from information in
the schema.

Secondly, this allowed me to discover a bug in handling `<bpmn:property
bpxe:flow="...">`. In the existing code I forgot to to handle the fact
that it is in a different namespace, and I was testing against target
namespace, which would have failed in real life. This change fixes this
issue.

This change also ensures values coming from these keyword lists will be
stringified (so that booleans or numbers) can be passed without quotes.

NB: this will not handle attribute values that are not converting to
strings, such as tuples (we currently use them in synthesized IDs).

Partially addresses #1