Problem: tests often fail on GitHub
This is happening because our testing is often time-sensitive (expecting
messages to appear or not appear within a certain amount of time).

Solution: execute tests sequentially on GitHub
in the hope that they will time out less.

This is done as a temporary measure before a better testing method is
implemented (see #4).