Problem: bpxe won't compile on OTP 23.1
It works on 23.0 according to the test matrix, but not on 23.1

Error looks like this:

```
** (CompileError) lib/bpxe/engine/process.ex: internal error in beam_ssa_opt;
crash reason: {case_clause,{'EXIT',{{badmatch,false},
                      [{beam_ssa_type,simplify,2,
                                      [{file,"beam_ssa_type.erl"},{line,904}]},
                       {beam_ssa_type,simplify,5,
                                      [{file,"beam_ssa_type.erl"},{line,716}]},
                       {beam_ssa_type,opt_is,8,
                                      [{file,"beam_ssa_type.erl"},{line,514}]},
                       {beam_ssa_type,opt_bs,8,
                                      [{file,"beam_ssa_type.erl"},{line,453}]},
                       {beam_ssa_type,opt_function,5,
                                      [{file,"beam_ssa_type.erl"},{line,431}]},
                       {beam_ssa_opt,ssa_opt_type_continue,1,
                                     [{file,"beam_ssa_opt.erl"},{line,441}]},
                       {compile,run_sub_passes_1,3,
                                [{file,"compile.erl"},{line,447}]},
                       {beam_ssa_opt,phase,4,
                                     [{file,"beam_ssa_opt.erl"},
                                      {line,114}]}]}}}
```

Solution: work around SSA optimizer

After an extended debugging session and using `dbg` call tracing, I was
able to catch the cause for this error.

I was calling `handle_call` internally in one of the `Process` module
functions, and `handle_call` can return either `{:reply, response,
state}` or `{:noreply, state}` (at least that's the union Erlang
compiler/optimizer deduced) and I was not handling the `:noreply`
variant.

`:beam_ssa_type.simplify/2` was called for `get_tuple_element` with an
index of 2, however there's an assertion in that function that says that
the size should be greater than the index (and it raises `badmatch` if
it fails). Between `:reply` and `:noreply`, they are of length 3 and 2,
respectively. A union, however, had a guaranteed length of 2 (at least
that's the way I understood what's happening) which was not greater than
the index of `2` (which was trying to access `state` from the `:reply`
variant).

So, on one hand, the compiler caught a potential crash (if there was a
`:noreply` response), on the other hand, it didn't make it easy to
figure out and understand.

I switched to using a more appropriate function (`add_flow_element`)
that does not return a variant of differently sized tuples.

I still suspect that this is a bug in optimizer; at least if it crashes,
it should be a but more... informative?